+++ runtime/near-vm-logic/src/gas_counter.rs
@@ -1,5 +1,4 @@
 use crate::{HostError, VMLogicError};
-use near_primitives::types::TrieNodesCount;
 use near_primitives_core::config::ExtCosts::read_cached_trie_node;
 use near_primitives_core::config::ExtCosts::touching_trie_node;
 use near_primitives_core::{
@@ -7,6 +6,7 @@ use near_primitives_core::{
     profile::ProfileDataV3,
     types::Gas,
 };
+use near_vm_errors::TrieNodesCount;
 use std::collections::HashMap;
 
 #[inline]
